+++ src/Konsole.Tests/WindowTests/WriteLineShould.cs
@@ -7,6 +7,21 @@ namespace Konsole.Tests.WindowTests
     class WriteLineShould
     {
         [Test]
+        public void when_writing_to_last_char_on_screen_move_cursor_forward_two_lines()
+        {
+            var con = new MockConsole(6, 4);
+            con.WriteLine("123456");
+            con.WriteLine("XY");
+            var expected = new[]
+            {
+                "123456",
+                "      ",
+                "XY    ",
+                "      "
+            };
+            con.Buffer.Should().BeEquivalentTo(expected);
+        }
+        [Test]
         public void allow_embedded_interpolations_without_exception()
         {
             var con = new MockConsole(6, 4);