+++ change-log.md
@@ -2,6 +2,13 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 
+## [6.0.0-alpha]
+
+### fixed
+
+- #40 WriteLine conflicts with String.Format 
+- #41 If a `Write` ends exactly on the last column, then the cursor is not advanced to the next line. 
+
 ## [5.4.4]
 
 ### Added